Rewrite code for 1.16.5

+++ src/main/java/power/keepeersofthestones/procedures/AccelerationTimeUseProcedure.java
@@ -84,7 +84,7 @@ public class AccelerationTimeUseProcedure {
 					((LivingEntity) entity)
 							.addPotionEffect(new EffectInstance(AccelerationTimeEffectPotionEffect.potion, (int) 400, (int) 0, (false), (false)));
 				if (world instanceof World && !world.isRemote()) {
-					((World) world).playSound(null, new BlockPos(x, y, z),
+					((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
 							(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("power:time_power")),
 							SoundCategory.PLAYERS, (float) 1, (float) 1);
 				} else {
